fix(link algorithm): refactored to use double instead of int + more error handling
fix(epsilon): now use argument instead of built-in variable

fix(GG): now doesn't create 2 time the same edges

+++ src/main/java/main/projectneighbourgraph/linkStrategy/LinkStrategy.java
@@ -16,7 +16,8 @@ public interface LinkStrategy {
      * @param nodeList list of nodes to link
      * @param arg Array of arguments to use in the link strategy
      * @param distanceStrategy distance strategy to use
+     * @return list of edges created
      */
-    ArrayList<Edge> link(ArrayList<Node> nodeList, int[] arg, DistanceStrategy distanceStrategy) throws Exception ;
+    ArrayList<Edge> link(ArrayList<Node> nodeList, double[] arg, DistanceStrategy distanceStrategy) throws Exception ;
 
 }
